Moved results components and container to their own folder
