+++ frontend/management/src/components/applicant-personal/results/SolvedAssignment.tsx
@@ -1,4 +1,4 @@
-import {AssignmentInterface} from "../../utils/types.tsx";
+import {AssignmentInterface} from "../../../utils/types.tsx";
 import {ReactNode} from "react";
 
 function SolvedAssignment({assignment}: Props): ReactNode {
